Server / Client - Various enhancements.
- Make it more clear when a terror attack is successful but the target
found out who did it.
- Report a numeric HDI in the country information window.
- Shortcut: Introduce a way of creating 10 covert cells at a time; if
the covert cell is named "-x" where x is a positive whole number, then
ten cells named x, x+1, x+2,..., x+9 will be created.
- In the covert actions window, remove the '%' sign from the security
level (in preparation for a new feature).
- Shortcut: When domestic consumption is toggled off for the Drugs
resource, then all public resources will be set to export all
production.
- Shortcut: When the Drugs resource is made private, then all resources
are set to private and taxed at 25-30%.
- Countries under all government types now demand the same amount of
resources for domestic consumption.
- Log out some data for each country when the game starts.
- Make the "population level" value continuous rather than discrete, and
have it uncapped.
- Scale the amount of military units the AI might build or buy based on
the country's revenue.
- Having more covert cells in a country now results in better mission
success rates.
- Tourism income only drops if the country is under attack on its own
territory, or if it's occupied by another country.
- Fix a defect so that better approval and stability lead to less
emigration, not more.
- Fix a defect where countries may be able to net export more than they
produce.
- Cap immigration at 2% instead of 1%, to match the emigration cap.

+++ Server/Include/sp2_data_access_layer.h
@@ -266,8 +266,8 @@ namespace SP2
       //! Sends all remembered actions since begin of iteration
       void SendHistory();
 
-		//! Find the population level, between 1 and 12, for a country
-		UINT32 PopulationLevel(UINT32 in_iCountryID) const;
+		//! Find the population level, starting from 1, for a country
+		REAL32 PopulationLevel(UINT32 in_iCountryID) const;
 
 		//! Return the total value (in $) of missiles for a specific country
 		REAL32 TotalMissileValue(UINT32 in_iCountryID) const;
